+++ Meta/Lagom/Fuzzers/FuzzShell.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <AK/StringView.h>
-#include <Shell/Shell.h>
+#include <LibShell/Shell.h>
 #include <stddef.h>
 #include <stdint.h>
