+++ Tests/CMakeLists.txt
@@ -19,6 +19,7 @@ add_subdirectory(LibMarkdown)
 add_subdirectory(LibPDF)
 add_subdirectory(LibRegex)
 add_subdirectory(LibSemVer)
+add_subdirectory(LibShell)
 add_subdirectory(LibSQL)
 add_subdirectory(LibTest)
 add_subdirectory(LibTextCodec)
