+++ Tests/LibShell/Tests/control-structure-as-command.sh
