+++ Tests/LibShell/Tests/sigpipe.sh
