+++ Userland/BuggieBox/main.cpp
@@ -7,7 +7,7 @@
 
 #include <AK/LexicalPath.h>
 #include <LibMain/Main.h>
-#include <Userland/Shell/Shell.h>
+#include <LibShell/Shell.h>
 
 #define ENUMERATE_UTILITIES(E, ALIAS) \
     ALIAS(b2sum, checksum)            \
