+++ Userland/Libraries/CMakeLists.txt
@@ -50,6 +50,7 @@ add_subdirectory(LibRegex)
 add_subdirectory(LibRIFF)
 add_subdirectory(LibSanitizer)
 add_subdirectory(LibSemVer)
+add_subdirectory(LibShell)
 add_subdirectory(LibSoftGPU)
 add_subdirectory(LibSQL)
 add_subdirectory(LibSymbolication)
