+++ Userland/Libraries/LibShell/Formatter.cpp
