+++ Userland/Libraries/LibShell/ImmediateFunctions.cpp
