+++ Userland/Libraries/LibShell/NodeVisitor.cpp
