+++ Userland/Libraries/LibShell/NodeVisitor.h
