+++ Userland/Libraries/LibShell/Parser.cpp
