+++ Userland/Libraries/LibShell/SyntaxHighlighter.h
